README.md
Name: Pac-Man chump exercise 
Description: This is an exercise that is for MIT's MERN stacking coding certificate.
Installation: All you need to get started with this project is a text editor. 
Support: I used StackOverflow and youtube to help troubleshoot difficulties.
Roadmap: I plan to add vertical movement to this project so that the Pac-Man image travels around to the entire perimeter of the webpage.